AgriFixAR_Python_Client/requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-generativeai
google-genai
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-text-splitters
chromadb
pypdf
python-dotenv
pydantic
typing-extensions
requests
Pillow
torch==2.2.2
transformers==4.41.2
opencv-python-headless==4.9.0.80